+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.8 (unreleased)
 
 - Change `Workspace` link on AnVIL to use `anvil.terra.bio` instead of `app.terra.bio`
+- Add ability to send a notification email when a user links their AnVIL account.
 
 ## 0.7 (2022-12-07)